+++ lib/image_optim/config.rb
@@ -161,7 +161,7 @@ class ImageOptim
       when true, nil
         {}
       when false
-        false
+        {:disable => true}
       else
         fail ConfigurationError, "Got #{worker_options.inspect} for "\
             "#{klass.name} options"